controller dao

+++ src/main/java/lucatic/grupo1/repository/DAOPerfil.java
@@ -2,7 +2,9 @@ package lucatic.grupo1.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DAOPerfil extends JpaRepository<Perfil, Integer>{
+import lucatic.grupo1.model.Perfil;
+
+public interface DAOPerfil extends JpaRepository<Perfil, Long>{
